requirements.txt
gunicorn==21.2.0
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
av==15.0.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.5
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
Django==5.2.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
faster-whisper==1.1.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.3
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-ollama==0.3.6
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.9
langsmith==0.4.11
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.5
mediapipe==0.10.21
ml_dtypes==0.5.3
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.3
numba==0.61.2
numpy==1.26.4
ollama==0.5.1
onnxruntime==1.22.1
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
portalocker==3.2.0
proglog==0.1.12
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.42
sqlparse==0.5.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.54.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
